basic custom email template editor, model and preview endpoint.
